Organizes custom maya class files into folders
